Move `naive_layout_of` query provider in separate sibling module
